Correcciones sobre solicitudes GET y DELETE

+++ src/products/products.service.ts
@@ -1,4 +1,10 @@
-import { BadRequestException, Injectable, InternalServerErrorException, Logger } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  InternalServerErrorException,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -7,16 +13,14 @@ import { Product } from './entities/product.entity';
 
 @Injectable()
 export class ProductsService {
-
   private readonly logger = new Logger();
 
   constructor(
     @InjectRepository(Product)
-    private readonly productRepository: Repository<Product>
+    private readonly productRepository: Repository<Product>,
   ) {}
 
   async create(createProductDto: CreateProductDto) {
-
     try {
       const product = this.productRepository.create(createProductDto);
       await this.productRepository.save(product);
@@ -26,6 +30,7 @@ export class ProductsService {
     }
   }
 
+  // TODO: Paginar
   async findAll() {
     try {
       const products = await this.productRepository.find();
@@ -36,31 +41,34 @@ export class ProductsService {
   }
 
   async findOne(id: string) {
-    try {
-      const product = await this.productRepository.findOneBy({
-        id: id
-      });
-      return product;
-    } catch (error) {
-      this.handleDBExceptions(error);
+    const product = await this.productRepository.findOneBy({
+      id: id,
+    });
+    if (!product) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
     }
+    return product;
   }
 
   update(id: number, updateProductDto: UpdateProductDto) {
     return `This action updates a #${id} product`;
   }
 
-  remove(id: string) {
-    try {
-      this.productRepository.delete({id: id});
-      return;
-    } catch (error) {
-      this.handleDBExceptions(error);
+  async remove(id: string) {
+    const product = await this.findOne(id);
+
+    if (!product) {
+      return new NotFoundException(`Product with ID ${id} not found`);
     }
+
+    this.productRepository.delete({id: id});
+
+    return;
   }
 
   private handleDBExceptions(error: any) {
-    if (error.code = '23505') {
+    console.log(`handleDBExceptions(${error})`);
+    if ((error.code = '23505')) {
       throw new BadRequestException(error.detail);
     }
